Clearer to use @TestExtension to keep all related code next to the test case.
